Made tweaks to the island validation test, after discussion with Ian in mid-December. Also fixed small 'bug' (species.coords attribute's values were being set using np.int32, which for floats very close to but below an integer (e.g. 34.9999845) was coercing them upward rather than down, so that they were then appearing to be in the wrong raster cells, such that their density-dependent death probabilities (which use spp.coords) disagreed with their extracted environmental values (which use ind.x, ind.y)).

+++ structs/species.py
@@ -597,7 +597,7 @@ class Species(OD):
             ig = itemgetter(*individs)
             coords = ig(dict(zip([*self], coords)))
         if as_float == True:
-            coords = np.float32(coords)
+            coords = np.float64(coords)
         else:
             coords = np.int32(np.floor(coords))
         #make sure it's at least 2d (in case a single individual is requested)
